fix(jupyter): use bind mount to allow cindex-ing
First tried this approach:
1. mount the codesearch index and the notebooks volume together into a new
   container,
2. run cindex over the notebooks.
But this failed as cindex didn't update the index. I guess because it
didn't have access to all indexed directories (I didn't mount the other
directories).

But it still wouldn't work either way. Even if I could update the index,
it wouldn't have access to the notebooks when running search on the
host.

The lesson? Index needs access to all data!